web_requirements.txt
# Web App Requirements for ISL Recognition
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
tensorflow==2.13.0
scikit-learn==1.3.0
gTTS==2.4.0
pygame==2.5.2
pyttsx3==2.90

# Additional web dependencies
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Audio processing
pyaudio==0.2.11

# For better security in production
python-dotenv==1.0.0